🗃️ refactor(execution): delete legacy service and consolidate to single execution.service.ts (#45)
- delete unused execution.service.ts (the original pre-configured version)
- rename configured-execution.service.ts → execution.service.ts
- update all imports in module, controller, and reconciliation service
- fix pre-existing TS2322 type error in PR-creation code (closes #46)
