+++ src/modules/execution/execution.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 import { GitHubModule } from "../github/github.module.js";
 import { GraphModule } from "../graph/graph.module.js";
 import { ExecutionController } from "./execution.controller.js";
-import { ExecutionService } from "./configured-execution.service.js";
+import { ExecutionService } from "./execution.service.js";
 
 @Module({
   imports: [GraphModule, GitHubModule],
