+++ src/modules/execution/execution.service.ts
@@ -6,18 +6,24 @@ import { execFileSync } from "node:child_process";
 import { join, resolve } from "node:path";
 import { getConfig } from "../../config.js";
 import { getDefaultWorkingBranch, listDefaultWorkingBranches } from "./default-working-branches.js";
+import { GitHubService } from "../github/github.service.js";
 import { GraphService } from "../graph/graph.service.js";
 import { WorkItemsService } from "../graph/work-items.service.js";
 import type { WorkItemRecord } from "../graph/types.js";
 import type {
+  CreateExecutionPullRequestDto,
+  CreateExecutionPullRequestResult,
   ExecutionDispatchGroupPreview,
   ExecutionDispatchNodePreview,
   ExecutionDispatchPreview,
+  ExecutionPullRequestRecord,
   ExecutionRunRecord,
   ExecutionSafetyCheck,
   ExecutionStatusEvent,
   LaunchExecutionRunDto,
   LaunchExecutionRunResult,
+  SyncMergedExecutionDto,
+  SyncMergedExecutionResult,
 } from "./types.js";
 
 @Injectable()
@@ -34,6 +40,7 @@ export class ExecutionService {
   constructor(
     @Inject(GraphService) private readonly graphService: GraphService,
     @Inject(WorkItemsService) private readonly workItemsService: WorkItemsService,
+    @Inject(GitHubService) private readonly githubService: GitHubService,
   ) {}
 
   stream(): Observable<MessageEvent> {
@@ -164,6 +171,136 @@ export class ExecutionService {
     });
 
     return { accepted: true, run };
+  }
+
+  async createPullRequest(input: CreateExecutionPullRequestDto): Promise<CreateExecutionPullRequestResult> {
+    const runId = input.run_id?.trim();
+    if (!runId) {
+      throw new BadRequestException("run_id is required");
+    }
+
+    const run = this.getRun(runId);
+    if (!run) {
+      throw new BadRequestException(`Unknown execution run: ${runId}`);
+    }
+    if (run.status !== "completed") {
+      throw new BadRequestException(`Execution run ${runId} must be completed before creating a pull request`);
+    }
+    if (run.pull_request) {
+      return { run, pull_request: run.pull_request };
+    }
+
+    const workItem = this.workItemsService.get(run.work_item_id);
+    const baseRef = input.base_ref?.trim() || run.base_ref || getDefaultWorkingBranch(workItem.repo);
+    const title = input.title?.trim() || this.buildPullRequestTitle(workItem);
+    const body = input.body?.trim() || this.buildPullRequestBody(run, workItem, baseRef);
+    const aheadCount = this.countCommitsAhead(run.worktree_path, baseRef, run.branch);
+    if (aheadCount === 0) {
+      throw new BadRequestException(`Branch ${run.branch} has no commits ahead of ${baseRef}; nothing is ready to open as a pull request`);
+    }
+
+    this.runGitIn(run.worktree_path, ["push", "--set-upstream", "origin", run.branch]);
+    this.runGhIn(run.worktree_path, [
+      "pr",
+      "create",
+      "--base",
+      baseRef,
+      "--head",
+      run.branch,
+      "--title",
+      title,
+      "--body-file",
+      "-",
+      ...(input.draft ? ["--draft"] : []),
+    ], body);
+
+    const pullRequest = this.readPullRequest(run.worktree_path, title, body);
+    const nextRun = this.updateRun(run.run_id, { pull_request: pullRequest }) ?? run;
+    mkdirSync(join(run.artifact_dir, "outputs"), { recursive: true });
+    writeFileSync(join(run.artifact_dir, "outputs", "pull-request.json"), JSON.stringify(pullRequest, null, 2), "utf8");
+    this.appendEvent(nextRun, { type: "pull_request_created", pull_request: pullRequest });
+    this.writeSummary(nextRun);
+
+    return { run: nextRun, pull_request: pullRequest };
+  }
+
+  async syncMergedPullRequest(input: SyncMergedExecutionDto): Promise<SyncMergedExecutionResult> {
+    const workItemId = input.work_item_id?.trim();
+    if (!workItemId) {
+      throw new BadRequestException("work_item_id is required");
+    }
+
+    const workItem = this.workItemsService.get(workItemId);
+    if (!workItem.repo) {
+      throw new BadRequestException(`Work item ${workItemId} is missing repo metadata required for PR sync`);
+    }
+
+    const pullRequest = input.pull_request_number
+      ? await this.githubService.readPullRequest(workItem.repo, input.pull_request_number)
+      : await this.resolvePullRequestFromWorkItem(workItem);
+
+    if (!pullRequest.merged_at) {
+      throw new BadRequestException(`Pull request #${pullRequest.number} has not been merged yet`);
+    }
+
+    const matchedRun = this.findRecentRunForSync(workItem, pullRequest.number);
+    const actualFilesSelection = this.selectActualFilesForMergeSync(input, workItem, matchedRun);
+    const nextBranch = this.normalizeNullableString(input.branch) ?? workItem.branch ?? (pullRequest.head_ref || null);
+    const nextArchivePath = this.hasOwn(input, "archive_path") ? (input.archive_path ?? null) : workItem.archive_path;
+    const nextMeta = this.buildMergedWorkItemMeta(workItem, pullRequest, matchedRun?.run_id ?? null, actualFilesSelection.source);
+
+    this.workItemsService.update(workItem.id, {
+      state: "done",
+      actual_files: actualFilesSelection.files,
+      branch: nextBranch,
+      archive_path: nextArchivePath,
+      meta: nextMeta,
+    });
+
+    const updatedWorkItem = this.workItemsService.get(workItem.id);
+    const managedBlockSync = input.stage_github_sync ? await this.safeStageManagedBlockSync(updatedWorkItem.id) : undefined;
+
+    if (matchedRun) {
+      const nextPullRequest: ExecutionPullRequestRecord = {
+        ...(matchedRun.pull_request ?? this.toExecutionPullRequestRecord(pullRequest)),
+        number: pullRequest.number,
+        url: pullRequest.url,
+        title: pullRequest.title,
+        body: pullRequest.body,
+        base_ref: pullRequest.base_ref,
+        head_ref: pullRequest.head_ref,
+        is_draft: pullRequest.is_draft,
+        state: pullRequest.state,
+        merged_at: pullRequest.merged_at,
+        merge_commit_sha: pullRequest.merge_commit_sha,
+      };
+      const syncedRun = this.updateRun(matchedRun.run_id, { pull_request: nextPullRequest }) ?? matchedRun;
+      this.appendEvent(syncedRun, {
+        type: "post_merge_sync_completed",
+        work_item_id: updatedWorkItem.id,
+        pull_request: nextPullRequest,
+        actual_files: actualFilesSelection.files,
+      });
+      this.writeSummary(syncedRun);
+    }
+
+    return {
+      synced: true,
+      work_item: {
+        id: updatedWorkItem.id,
+        state: updatedWorkItem.state,
+        branch: updatedWorkItem.branch,
+        archive_path: updatedWorkItem.archive_path,
+        actual_files: updatedWorkItem.actual_files,
+        meta: updatedWorkItem.meta,
+        updated_at: updatedWorkItem.updated_at,
+      },
+      pull_request: this.toExecutionPullRequestRecord(pullRequest),
+      matched_run_id: matchedRun?.run_id ?? null,
+      actual_files_source: actualFilesSelection.source,
+      dispatch_preview: this.getPreview(updatedWorkItem.repo ?? undefined),
+      managed_block_sync: managedBlockSync,
+    };
   }
 
   private async executeRun(initialRun: ExecutionRunRecord, node: ExecutionDispatchNodePreview) {
@@ -327,6 +464,7 @@ export class ExecutionService {
       `Issue URL: ${workItem.issue_url ?? "(not set)"}`,
       `Scope hint: ${workItem.scope_hint ?? "(not set)"}`,
       `Suggested execution branch: ${node.branch}`,
+      `Default working base branch: ${node.default_base_ref}`,
       "",
       "Files owned:",
       owned,
@@ -451,6 +589,7 @@ export class ExecutionService {
       `- Updated: ${run.updated_at}`,
       ...(run.started_at ? [`- Started: ${run.started_at}`] : []),
       ...(run.completed_at ? [`- Completed: ${run.completed_at}`] : []),
+      ...(run.pull_request ? [`- Pull request: ${run.pull_request.url}`] : []),
       "",
       "## Safety checks",
       ...run.safety_checks.map((check) => `- [${check.status}] ${check.code}: ${check.message}`),
@@ -460,6 +599,15 @@ export class ExecutionService {
       run.result_summary ?? run.progress_message ?? "No summary available.",
       "",
       ...(run.changed_files?.length ? ["## Changed files", ...run.changed_files.map((path) => `- ${path}`), ""] : []),
+      ...(run.pull_request ? [
+        "## Pull request",
+        `- Number: ${run.pull_request.number}`,
+        `- URL: ${run.pull_request.url}`,
+        `- Draft: ${run.pull_request.is_draft ? "yes" : "no"}`,
+        `- Base: ${run.pull_request.base_ref}`,
+        `- Head: ${run.pull_request.head_ref}`,
+        "",
+      ] : []),
     ];
 
     writeFileSync(join(run.artifact_dir, "summary.md"), lines.join("\n"), "utf8");
@@ -488,8 +636,25 @@ export class ExecutionService {
   }
 
   private runGit(args: string[], options: { allowFailure?: boolean } = {}): string {
+    return this.runCommand("git", args, { cwd: process.cwd(), allowFailure: options.allowFailure });
+  }
+
+  private runGitIn(cwd: string, args: string[], options: { allowFailure?: boolean } = {}): string {
+    return this.runCommand("git", args, { cwd, allowFailure: options.allowFailure });
+  }
+
+  private runGhIn(cwd: string, args: string[], stdin?: string): string {
+    return this.runCommand("gh", args, { cwd, stdin });
+  }
+
+  private runCommand(command: string, args: string[], options: { cwd: string; stdin?: string; allowFailure?: boolean }): string {
     try {
-      return execFileSync("git", args, { cwd: process.cwd(), encoding: "utf8" });
+      return execFileSync(command, args, {
+        cwd: options.cwd,
+        input: options.stdin,
+        encoding: "utf8",
+        maxBuffer: 1024 * 1024 * 4,
+      });
     } catch (error) {
       if (options.allowFailure) {
         return "";
@@ -521,6 +686,73 @@ export class ExecutionService {
     return path.trim();
   }
 
+  private countCommitsAhead(worktreePath: string, baseRef: string, branch: string): number {
+    const output = this.runGitIn(worktreePath, ["rev-list", "--count", `${baseRef}..${branch}`], { allowFailure: true }).trim();
+    const parsed = Number(output);
+    return Number.isFinite(parsed) ? parsed : 0;
+  }
+
+  private readPullRequest(worktreePath: string, fallbackTitle: string, body: string): ExecutionPullRequestRecord {
+    const raw = this.runGhIn(worktreePath, ["pr", "view", "--json", "number,url,title,body,baseRefName,headRefName,isDraft"]);
+    const parsed = JSON.parse(raw) as {
+      number?: number;
+      url?: string;
+      title?: string;
+      body?: string;
+      baseRefName?: string;
+      headRefName?: string;
+      isDraft?: boolean;
+    };
+
+    if (!Number.isInteger(parsed.number) || !parsed.url || !parsed.baseRefName || !parsed.headRefName) {
+      throw new BadRequestException("Created pull request could not be read back from GitHub safely");
+    }
+
+    return {
+      number: parsed.number!,
+      url: parsed.url,
+      title: parsed.title?.trim() || fallbackTitle,
+      body: parsed.body ?? body,
+      base_ref: parsed.baseRefName!,
+      head_ref: parsed.headRefName!,
+      is_draft: Boolean(parsed.isDraft),
+      created_at: this.now(),
+    };
+  }
+
+  private buildPullRequestTitle(workItem: WorkItemRecord): string {
+    const issuePrefix = workItem.issue_number ? `[#${workItem.issue_number}] ` : "";
+    return `${issuePrefix}${workItem.name}`;
+  }
+
+  private buildPullRequestBody(run: ExecutionRunRecord, workItem: WorkItemRecord, baseRef: string): string {
+    const lines = [
+      "## Summary",
+      run.result_summary ?? run.progress_message ?? `Completed Studio execution run ${run.run_id}.`,
+      "",
+      "## Context",
+      `- Work item: ${workItem.id}`,
+      `- Issue: ${workItem.issue_url ?? "(not linked)"}`,
+      `- Base branch: ${baseRef}`,
+      `- Head branch: ${run.branch}`,
+      `- Artifact dir: ${run.artifact_dir}`,
+      `- Worktree: ${run.worktree_path}`,
+    ];
+
+    if (workItem.scope_hint) {
+      lines.push(`- Scope hint: ${workItem.scope_hint}`);
+    }
+
+    lines.push("", "## Changed files");
+    if (run.changed_files?.length) {
+      lines.push(...run.changed_files.map((path) => `- \`${path}\``));
+    } else {
+      lines.push("- (not recorded)");
+    }
+
+    return lines.join("\n");
+  }
+
   private syncActualFiles(workItemId: string, changedFiles: string[]): { ok: true; actual_files: string[] } | { ok: false; message: string } {
     try {
       const actualFiles = [...new Set(changedFiles.filter(Boolean))].sort();
@@ -530,6 +762,144 @@ export class ExecutionService {
       this.logger.warn(`Failed to sync actual_files for ${workItemId}: ${this.getErrorMessage(error)}`);
       return { ok: false, message: this.getErrorMessage(error) };
     }
+  }
+
+  private async resolvePullRequestFromWorkItem(workItem: WorkItemRecord) {
+    const branch = workItem.branch?.trim();
+    if (!branch) {
+      throw new BadRequestException(`Work item ${workItem.id} is missing branch metadata required to resolve its pull request`);
+    }
+
+    const pullRequest = await this.githubService.findPullRequestForBranch(workItem.repo ?? "", branch);
+    if (!pullRequest) {
+      throw new BadRequestException(`No pull request was found for branch ${branch}`);
+    }
+
+    return pullRequest;
+  }
+
+  private findRecentRunForSync(workItem: WorkItemRecord, pullRequestNumber: number): ExecutionRunRecord | null {
+    return this.listRecentRuns().find((run) => {
+      if (run.work_item_id !== workItem.id) {
+        return false;
+      }
+      if (run.pull_request?.number === pullRequestNumber) {
+        return true;
+      }
+      return run.branch === workItem.branch;
+    }) ?? null;
+  }
+
+  private selectActualFilesForMergeSync(
+    input: SyncMergedExecutionDto,
+    workItem: WorkItemRecord,
+    matchedRun: ExecutionRunRecord | null,
+  ): { files: string[]; source: "input" | "work_item" | "run" } {
+    if (Array.isArray(input.actual_files) && input.actual_files.length > 0) {
+      return { files: this.uniqueSorted(input.actual_files), source: "input" };
+    }
+    if (workItem.actual_files.length > 0) {
+      return { files: this.uniqueSorted(workItem.actual_files), source: "work_item" };
+    }
+    if (matchedRun?.changed_files?.length) {
+      return { files: this.uniqueSorted(matchedRun.changed_files), source: "run" };
+    }
+    return { files: [], source: "work_item" };
+  }
+
+  private buildMergedWorkItemMeta(
+    workItem: WorkItemRecord,
+    pullRequest: {
+      number: number;
+      url: string;
+      title: string;
+      state: string;
+      base_ref: string;
+      head_ref: string;
+      merged_at: string | null;
+      merge_commit_sha: string | null;
+    },
+    runId: string | null,
+    actualFilesSource: "input" | "work_item" | "run",
+  ): Record<string, unknown> {
+    return {
+      ...workItem.meta,
+      studio_post_merge_sync: {
+        synced_at: this.now(),
+        run_id: runId,
+        actual_files_source: actualFilesSource,
+        pull_request: {
+          number: pullRequest.number,
+          url: pullRequest.url,
+          title: pullRequest.title,
+          state: pullRequest.state,
+          base_ref: pullRequest.base_ref,
+          head_ref: pullRequest.head_ref,
+          merged_at: pullRequest.merged_at,
+          merge_commit_sha: pullRequest.merge_commit_sha,
+        },
+      },
+    };
+  }
+
+  private async safeStageManagedBlockSync(workItemId: string) {
+    try {
+      const staged = await this.githubService.stageManagedBlockSync(workItemId);
+      return {
+        work_item_id: staged.work_item_id,
+        based_on_body_hash: staged.based_on_body_hash,
+        operations: staged.operations,
+      };
+    } catch (error) {
+      this.logger.warn(`Failed to stage managed block sync for ${workItemId}: ${this.getErrorMessage(error)}`);
+      return null;
+    }
+  }
+
+  private toExecutionPullRequestRecord(pullRequest: {
+    number: number;
+    url: string;
+    title: string;
+    body: string;
+    base_ref: string;
+    head_ref: string;
+    is_draft: boolean;
+    state: string;
+    merged_at: string | null;
+    merge_commit_sha: string | null;
+  }): ExecutionPullRequestRecord {
+    return {
+      number: pullRequest.number,
+      url: pullRequest.url,
+      title: pullRequest.title,
+      body: pullRequest.body,
+      base_ref: pullRequest.base_ref,
+      head_ref: pullRequest.head_ref,
+      is_draft: pullRequest.is_draft,
+      created_at: this.now(),
+      state: pullRequest.state,
+      merged_at: pullRequest.merged_at,
+      merge_commit_sha: pullRequest.merge_commit_sha,
+    };
+  }
+
+  private normalizeNullableString(value?: string | null): string | null | undefined {
+    if (typeof value === "undefined") {
+      return undefined;
+    }
+    if (value === null) {
+      return null;
+    }
+    const trimmed = value.trim();
+    return trimmed.length > 0 ? trimmed : null;
+  }
+
+  private hasOwn<T extends object>(value: T, key: PropertyKey): boolean {
+    return Object.prototype.hasOwnProperty.call(value, key);
+  }
+
+  private uniqueSorted(values: string[]): string[] {
+    return [...new Set(values.map((value) => value.trim()).filter(Boolean))].sort();
   }
 
   private getWorktreePath(branch: string): string {
